backend/Idempotency-gateway/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.2
PyJWT==2.12.1
python-dotenv==1.2.2
redis==7.4.0
sqlparse==0.5.5
whitenoise==6.7.0
psycopg2-binary==2.9.10
